feat(): [FOR] Reorganize directives
