+++ src/main/resources/public/ts/directives/index.ts
@@ -1,7 +1,4 @@
 export * from './formulaire-picker-file';
 export * from './infinite-scroll';
 export * from './loader';
-export * from './question-item';
-export * from './question-title';
-export * from './question-type';
-export * from './question-type/index';
+export * from './question/index';
