+++ src/main/resources/public/ts/directives/question/index.ts
@@ -1,0 +1,4 @@
+export * from './question-item';
+export * from './question-title';
+export * from './question-type';
+export * from './question-type/index';
