+++ src/main/resources/public/ts/directives/question/question-type.ts
@@ -1,5 +1,5 @@
 import {Directive, ng} from "entcore";
-import {Question, Types} from "../models";
+import {Question, Types} from "../../models";
 
 interface IViewModel {
     question: Question,
